requirements.txt
aiogram>=3.4.1
pydantic-settings>=2.2.1
openai>=1.12.0
python-dotenv>=1.0.1
fastapi>=0.109.0
uvicorn>=0.27.0
